+++ src/test/java/it/pagopa/pn/bff/pnclient/apikeys/PnPublicKeyManagerClientPGImplTest.java
@@ -1,4 +1,4 @@
-package it.pagopa.pn.bff.pnclient.publickeys;
+package it.pagopa.pn.bff.pnclient.apikeys;
 
 import it.pagopa.pn.bff.generated.openapi.msclient.publickey_pg.api.PublicKeysApi;
 import it.pagopa.pn.bff.generated.openapi.msclient.publickey_pg.model.CxTypeAuthFleet;
